movies_api/app/requirements.txt
elasticsearch[async]==8.13.2
fastapi==0.111.0
pydantic==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
aioredis==2.0.1
redis==5.0.4
pydantic_settings==2.4.0
opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-instrumentation-fastapi==0.38b0
opentelemetry-exporter-jaeger==1.17.0